+++ .vscode/settings.json
@@ -7,5 +7,6 @@
         "out": true // set this to false to include "out" folder in search results
     },
     "typescript.tsdk": "./node_modules/typescript/lib", // we want to use the TS server from our node_modules folder to control its version
-    "eslint.enable": false
+    "eslint.enable": false,
+    "tslint.autoFixOnSave": true
 }
