+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
     "compilerOptions": {
+        "alwaysStrict": true,
         "module": "commonjs",
         "target": "es6",
         "outDir": "out",
@@ -11,7 +12,6 @@
     },
     "exclude": [
         "node_modules",
-        ".vscode-test",
-        "_jest-editor"
+        ".vscode-test"
     ]
 }
