+++ tslint.json
@@ -3,6 +3,11 @@
 		"no-unused-expression": true,
 		"no-duplicate-variable": true,
 		"no-unused-variable": true,
+		"no-var-keyword": true,
+		"quotemark": [
+			true,
+			"single"
+		],
 		"curly": true,
 		"class-name": true,
 		"semicolon": ["never"],
